Separate the posix part of the c-ares driver
